main-api/requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-login==0.6.3
flask-cors==5.0.0
werkzeug==3.1.3
python-dotenv==1.0.1
requests==2.32.3
anthropic>=0.40.0
mistune==3.0.2
python-slugify==8.0.4
caldav==1.3.9
icalendar==6.1.1
beautifulsoup4>=4.12.0
lxml>=5.0.0
matplotlib>=3.8.0
numpy>=1.26.0
scipy>=1.12.0
scikit-learn>=1.4.0
